+++ lua/adev/custom-plugins/example.lua
@@ -1,0 +1,6 @@
+-- NOTE: do not modify this file, it will be overwritten on update
+-- create a new file in adev/custom-plugins if you want to add plugins
+-- This is an example of a custom plugin
+-- You can add your own plugins here
+
+return {}